using a keymap to mange controls
the control subsystem was completely rewritten. The control class was moved to the control namespace and renamed to input_handler. In addition to that it is no longer a singleton, which makes the class even cleaner. The latest version of the simulator is required for this to work. These changes break the simulator.