started rework of the video component
there is now a generic renderer and all of the ffmpeg code was moved into a child of that generic renderer.
The ffmpeg_sdl backend is now very small thanks to those changes. The next thing to come is a generic window management and adapting the opencv backend to this new structure. This might allow opencv/ffmpeg and more alternatives to sdl2