rework of generic video backend
Now the rework of the video backend is almost finished. The generic video backend now uses both the generic window manager and generic renderer. This allows a very flexible structure and enables developers to implement their own window manager and by building it on top of these classes.

Next the classes will be moved around in the namespaces and some documentation will be added.