reworks of the way backends are defined
before there was an enum which contained a list of all of the possible backends. This is now replaced by a class and some constant objects. This way new backends can easily be defined and other parts of the program can figure out which rendererType and which WindowType is used. An example of this is in the control system. The control system only cares about the window type and can completely ignore the renderer. Because of this no change is needed if a different renderer is used for as long as it is one of the supported windowing systems.
Everything in the video module should now be in the Video namespace except for the video singleton.