rework of the frame handling
frames are now only passed as shared_ptr. This new design fixes the memory problem mentioned in #17. In addition to that there is no longer a need to worry about clearing the frames, since this is done automatically in the destructor. This is possible since teh destructor only gets called, when the last reference to that address is deleted, so the