some changes to framebuffer textures to have more control
Now there is an option to get teh FBO variable from a framebuffer texture. This allows more fine control over how the framebuffer is loaded in cases where this is needed.  Also the default framebuffer clear color is set to black again since it can cause problem if it is something else.