added common opengl parts to imgui handler
Now there is the option to load a copy shader and a vertex array object as part of the imgui handler. The copy shader is needed all of the time and the basic vertex array contains the verticies for a rectangle including the texture coordinates. This will most likely be the only VAO that is needed throughout the library.