change the way rendertimes are calculated
now when more than one frame has to be advanced the sum of the times for all of the individual frames is the rendertime. now only the the total time will be displayed instead of the individual times. This should reflect the actual times more closely since the individual frame will not be displayed.