queueing next to draw snowflakes in advance
With this commit more than one snowflake can be in queue to be drawn. This allows for multiple snowflakes to be drawn within one frame. While doing this the scaling of the spawn rate is now corrected/improved. Also a weather intensity of 1 is now the default. Higher values allow for more intense snow, while lower values lower the spawn rate. This gives a simulator implementation control over the strength of the snow