using a timestamp in seek (for now) to kind of fix it
Now in the seek function a timestamp is used to seek the frame since seeking a frame seems not to work. This does jump to an incorrect frame when the framerate is not constant but that is way better than the seek function crashing the decode and jumping to the start of the video.